requirements.txt
streamlit
gpxpy
Pillow
folium